+++ ext/markly/config.h
@@ -5,9 +5,7 @@
 extern "C" {
 #endif
 
-#ifndef HAVE_STDBOOL_H
 #define HAVE_STDBOOL_H
-#endif
 
 #ifdef HAVE_STDBOOL_H
   #include <stdbool.h>
@@ -70,8 +68,6 @@ CMARK_INLINE int c99_snprintf(char *outBuf, size_t size, const char *format, ...
 }
 
 #endif
-
-#define CMARK_DEBUG_NODES 0
 
 #ifdef __cplusplus
 }