+++ ext/markly/iterator.c
@@ -139,9 +139,14 @@ void cmark_node_own(cmark_node *root) {
       switch (cur->type) {
       case CMARK_NODE_TEXT:
       case CMARK_NODE_HTML_INLINE:
-      case CMARK_NODE_CODE:
       case CMARK_NODE_HTML_BLOCK:
         cmark_chunk_to_cstr(iter->mem, &cur->as.literal);
+        break;
+      case CMARK_NODE_CODE:
+        cmark_chunk_to_cstr(iter->mem, &cur->as.code.literal);
+        if (cur->as.code.info.len > 0) {
+          cmark_chunk_to_cstr(iter->mem, &cur->as.code.info);
+        }
         break;
       case CMARK_NODE_LINK:
         cmark_chunk_to_cstr(iter->mem, &cur->as.link.url);